debugging models

+++ segnlp/metrics/default_token_metric.py
@@ -10,10 +10,11 @@ from .metric_utils import f1_precision_recall
 def default_token_metric(df:pd.DataFrame, task_labels:dict):
 
     collected_scores = {}
+    
     for task in task_labels.keys():
 
-        targets = df.loc["TARGET", task].to_numpy()
-        preds = df.loc["PRED", task].to_numpy()
+        targets = df.loc["TARGET", task].to_numpy().astype(int)
+        preds = df.loc["PRED", task].to_numpy().astype(int)
 
         scores = f1_precision_recall(
                             targets = targets,
